pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "TopoBench"
dynamic = ["version"]
authors = [
    {name = "Topological Intelligence Team Authors", email = "tlscabinet@gmail.com"}
]
readme = "README.md"
description = "Topological Deep Learning"
license = {file = "LICENSE.txt"}
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11"
]
# STRICTLY PYTHON 3.11
requires-python = ">=3.11, <3.12"

dependencies=[
    "torch==2.3.0",
    "tqdm",
    "charset-normalizer",
    "numpy",
    "scipy",
    "requests",
    "scikit-learn",
    "matplotlib",
    "networkx",
    "pandas",
    "decorator",
    "hypernetx < 2.0.0",
    "trimesh",
    "spharapy",
    "hydra-core==1.3.2",
    "hydra-colorlog==1.2.0",
    "hydra-optuna-sweeper==1.2.0",
    "wandb",
    "tensorboard",
    "einops==0.7.0",
    "tabulate",
    "ipykernel",
    "notebook",
    "jupyterlab",
    "rich",
    "ogb",
    "rootutils",
    "topomodelx @ git+https://github.com/pyt-team/TopoModelX.git",
    "toponetx @ git+https://github.com/pyt-team/TopoNetX.git@c378925",
    "lightning==2.4.0",
    "torch-scatter",
    "torch-sparse",
    "torch-cluster",
]

[project.optional-dependencies]
doc = [
    "jupyter",
    "nbsphinx",
    "nbsphinx_link",
    "numpydoc",
    "sphinx>=7.0,<9.0",   # Sphinx 9 is incompatible with the required docutils
    "docutils<0.21",      # Required for nbsphinx-link
    "sphinx_gallery",
    "pydata-sphinx-theme",
    "myst_parser"
]
lint = [
    "pre-commit",
    "ruff"
]
test = [
    "pytest",
    "pytest-cov",
    "coverage",
    "jupyter",
    "mypy",
    "pytest-mock"
]

dev = ["TopoBench[test, lint]"]
all = ["TopoBench[dev, doc]"]

[project.urls]
homepage="https://geometric-intelligence.github.io/topobench/index.html"
repository="https://github.com/geometric-intelligence/TopoBench"

# ==============================================================================
#  UV CONFIGURATION
# ==============================================================================

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

# Default find-links (will be overwritten by bash script)
[tool.uv]
find-links = ["https://data.pyg.org/whl/torch-2.3.0+cu121.html"]

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu", marker = "sys_platform == 'darwin' or sys_platform == 'win32'" },
    { index = "pytorch-cu121", marker = "sys_platform == 'linux'" },
]

[tool.uv.extra-build-dependencies]
torch-cluster = ["torch==2.3.0"]
torch-scatter = ["torch==2.3.0"]
torch-sparse = ["torch==2.3.0"]

# ==============================================================================
#  TOOL CONFIGS
# ==============================================================================
[tool.black]
line-length = 79
target-version = ['py311']

[tool.docformatter]
wrap-summaries = 79
wrap-descriptions = 79

[tool.ruff]
target-version = "py311"
extend-exclude = ["test", "tutorials", "notebooks"]
line-length = 79

[tool.ruff.format]
docstring-code-format = false

[tool.ruff.lint]
select = ["F", "E", "W", "I", "UP", "B", "PIE", "Q", "RET", "SIM", "NPY", "PERF"]
fixable = ["ALL"]
ignore = ["E501", "RET504", "RET505", "NPY002", "UP038", "W293", "PERF203"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F403"]
"tests/*" = ["D"]

[tool.setuptools.dynamic]
version = {attr = "topobench.__version__"}

[tool.setuptools.packages.find]
include = ["topobench", "topobench.*"]

[tool.mypy]
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
disable_error_code = ["import-untyped"]
plugins = "numpy.typing.mypy_plugin"

[[tool.mypy.overrides]]
module = ["torch_cluster.*","networkx.*","scipy.spatial","scipy.sparse","toponetx.classes.simplicial_complex"]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "--capture=no"
pythonpath = ["."]

[tool.numpydoc_validation]
checks = ["all", "GL01", "ES01", "EX01", "SA01"]
exclude = ['\.undocumented_method$', '\.__init__$', '\.__repr__$']
